+++ Source/Shaders/BillboardCollectionVS.glsl
@@ -16,8 +16,6 @@ attribute vec4 pickColor;
 attribute vec4 color;
 #endif
 
-const vec2 czm_highResolutionSnapScale = vec2(1.0, 1.0);    // TODO
-
 varying vec2 v_textureCoordinates;
 
 #ifdef RENDER_FOR_PICK
@@ -106,7 +104,7 @@ void main()
 
     vec4 positionWC = czm_eyeToWindowCoordinates(positionEC);
     
-    vec2 halfSize = imageSize * scale * czm_highResolutionSnapScale;
+    vec2 halfSize = imageSize * scale * czm_resolutionScale;
     halfSize *= ((direction * 2.0) - 1.0);
     
     positionWC.xy += (origin * abs(halfSize));
@@ -136,7 +134,7 @@ void main()
     
     positionWC.xy += halfSize;
     positionWC.xy += translate;
-    positionWC.xy += (pixelOffset * czm_highResolutionSnapScale);
+    positionWC.xy += (pixelOffset * czm_resolutionScale);
 
     gl_Position = czm_viewportOrthographic * vec4(positionWC.xy, -positionWC.z, 1.0);
     v_textureCoordinates = textureCoordinates;